+++ publish/CHANGELOG.md
@@ -1,5 +1,6 @@
 - v1.107
   - Adds new field `admin` to the permissions.yaml to support character specific admin status.
+  - Adds support for permissions for all characters with the same host name (character id is optional).
   - Fixes devcommands status not being used (admin status always enabled cheat access).
   - Fixes `move_spawn` command not working for custom spawn locations.
   - Fixes `move_spawn` command causing error if no spawn was found (now attempts to make a new one).